+++ example/interrupts-per-second/interrupts-per-second.js
@@ -1,6 +1,7 @@
 var Epoll = require('../../build/Release/epoll').Epoll,
   fs = require('fs'),
-  valuefd = fs.openSync('/sys/class/gpio/gpio38/value', 'r+'), // 17
+  inputfd = fs.openSync('/sys/class/gpio/gpio7/value', 'r+'),
+  outputfd = fs.openSync('/sys/class/gpio/gpio8/value', 'r+'),
   value = new Buffer(1),  // The three Buffers here are global
   zero = new Buffer('0'), // to improve performance.
   one = new Buffer('1'),
@@ -14,19 +15,19 @@ var poller = new Epoll(function (err, fd, events) {
   count++;
 
   // Read GPIO value file. Reading also clears the interrupt.
-  fs.readSync(fd, value, 0, 1, 0);
+  fs.readSync(inputfd, value, 0, 1, 0);
 
   // Toggle GPIO value. This will eventually result
   // in the next interrupt being triggered.
   nextValue = value[0] === zero[0] ? one : zero;
-  fs.writeSync(fd, nextValue, 0, nextValue.length, 0);
+  fs.writeSync(outputfd, nextValue, 0, nextValue.length, 0);
 });
 
 time = process.hrtime(); // Get start time.
 
 // Start watching for interrupts. This will trigger the first interrupt
 // as the value file already has data waiting for a read.
-poller.add(valuefd, Epoll.EPOLLPRI);
+poller.add(inputfd, Epoll.EPOLLPRI);
 
 // Print interrupt rate to console after 5 seconds.
 setTimeout(function () {
@@ -37,6 +38,6 @@ setTimeout(function () {
   console.log(rate + ' interrupts per second');
 
   // Stop watching.
-  poller.remove(valuefd).close();
+  poller.remove(inputfd).close();
 }, 5000);